% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind}
\alias{bind}
\title{bind together two testargs objects}
\usage{
bind(object1, object2)
}
\arguments{
\item{object1}{object of class \code{testargs}}

\item{object2}{object of class \code{testargs}}
}
\description{
bind together two testargs objects
}
