% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c,testargs-method}
\alias{c,testargs-method}
\title{Bind together testarg objects}
\usage{
\S4method{c}{testargs}(x, ...)
}
\arguments{
\item{x}{object of class \code{testargs}}

\item{...}{objects of class \code{testargs} to be combined with \code{x}}
}
\description{
When the objects to be combined do not have the same arguments or the same
diagnostics, the combined object simply takes the union of the argument and
diagnostic names, and uses \code{rbind.fill} to combine the diagnostic data.
}
