% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{testargs-class}
\alias{testargs-class}
\title{\code{'testargs'} class}
\description{
This is the central class definition of the \code{testarguments} package, containing all information from a call to \code{\link{test_arguments}}
}
\section{Slots}{

\describe{
\item{\code{diagnostics_df}}{a \code{data.frame} containing the diagnostics for each combination of the supplied arguments}

\item{\code{arg_names}}{the argument names}

\item{\code{diagnostic_names}}{the diagnostic names}
}}

