% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{c,testargs-method}
\alias{c,testargs-method}
\title{Bind together testarg objects}
\usage{
\S4method{c}{testargs}(x, ...)
}
\arguments{
\item{x}{object of class \code{testargs}}

\item{...}{objects of class \code{testargs} to be combined with \code{x}}
}
\description{
Bind together testarg objects
}
