% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_arguments.R
\name{optimal_arguments}
\alias{optimal_arguments}
\title{Find the optimal argument combinations for each diagnostic}
\usage{
optimal_arguments(object, optimality_criterion = which.min)
}
\arguments{
\item{object}{an object of class \code{testargs}}

\item{optimality_criterion}{a function (or list of functions) that defines the optimality criterion for each diagnostic. Each function should return a single integer indicating the index of the optimal argument combination}
}
\description{
The measure of "best performer" is diagnostic dependent; for example,
we typically wish to minimise the RMSE and run time, but we want coverage
to be as close to the purported value as possible. Hence,
\code{optimal_arguments()} allows one to set the optimality criterion
individually.
}
\examples{
## See the example in ?test_diagnostics for this functions intended use
}
