% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_arguments.R
\name{optimal_arguments}
\alias{optimal_arguments}
\title{Find the optimal argument combinations for each diagnostic}
\usage{
optimal_arguments(object, optimality_criterion = which.min)
}
\arguments{
\item{object}{an object of class \code{testargs}}

\item{optimality_criterion}{a function (or list of functions) that defines the optimality criterion for each diagnostic. Each function should return a single integer indicating the index of the optimal argument combination}
}
\description{
Find the optimal argument combinations for each diagnostic
}
