% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_arguments.R
\name{optimal_arguments}
\alias{optimal_arguments}
\title{Find the optimal arguments fir each diagnostic}
\usage{
optimal_arguments(object, optimality_criterion = which.min)
}
\arguments{
\item{object}{an object of class testargs}

\item{optimality_criterion}{a function (or list of functions) that defines the optimality criterion for each diagnostic. Should return a single index}
}
\description{
Find the optimal arguments fir each diagnostic
}
