% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Visualise diagnostics across the tested arguments.}
\usage{
plot_diagnostics(
  object,
  focused_args = NULL,
  average_out_non_focused_args = TRUE,
  plot_order = NULL
)
}
\arguments{
\item{object}{an object of class \code{testargs} from a call to \code{test_arguments()}}

\item{focused_args}{the arguments we wish to plot. If \code{NULL} (default), all arguments are plotted}

\item{average_out_non_focused_args}{logical indicating whether we should average over the non-focused arguments}

\item{plot_order}{integer vector giving the order in which we are to assign arguments to the various aesthetics. If \code{plot_order = NULL} (default), the arguments are assigned based on their type (in the order 'numeric', 'integer', 'factor', 'character', 'logical'); otherwise, if \code{focused_args = NULL}, \code{plot_order} should be the same length as \code{object@arg_names}, and if \code{focused_args} is not \code{NULL}, \code{plot_order} should be the same length as \code{focused_args}}
}
\value{
a \code{ggplot} object
}
\description{
The idea is to make a faceted plot, where:
\itemize{
\item{The columns of the facet correspond to the diagnostic (e.g., RMSPE, CRPS, Time, etc.)}
\item{The y-axis corresponds to the value of the diagnostic scores}
\item{The x-axis corresponds to the values of the first argument}
\item{The colour scale and grouping correspond to the second argument (if present)}
\item{If a third argument is present, \code{facet_grid} is used, whereby columns correspond to levels of the third argument, and rows correspond to diagnostics. Note that \code{facet_grid} forces a given row to share a common y-scale, so the plot would be misleading if diagnostics were kept as columns}
\item{The shape of the points correspond to values of the fourth argument (if present)}
}
}
\examples{
## See the example in ?test_diagnostics for this functions intended use
}
\seealso{
\code{\link{test_arguments}}
}
