% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Visualise diagnostics across the tested arguments.}
\usage{
plot_diagnostics(
  diagnostics_df,
  arg_names,
  focused_args = NULL,
  average_out_non_focused_args = TRUE
)
}
\arguments{
\item{arg_names}{the columns which correspond to argument names in \code{diagnostic_df}}

\item{focused_args}{the arguments we wish to plot. If \code{NULL} (default), all arguments are plotted}

\item{average_out_non_focused_args}{logical indicating whether we should average over the non-focused arguments}

\item{diagnostic_df}{a \code{data.frame} from a call to \code{test_arguments}}
}
\value{
a \code{ggplot} object
}
\description{
The idea is to make a faceted plot, where:
\itemize{
\item{The columns of the facet correspond to the diagnostic (e.g., RMSPE, CRPS, Time, etc.)}
\item{The y-axis is the value of the given diagnostic}
\item{The x-axis is the value of the first argument we are optimising}
\item{The colour scale and grouping is the value of the second argument (if present)}
\item{The rows of the facet correspond to the third argument (if present)}
\item{The shape of the points correspond to the fourth argument (if present)}
}
}
\examples{
## See the example in ?test_diagnostics for this functions intended use
}
\seealso{
\code{\link{test_arguments}}
}
