% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_arguments.R
\name{test_arguments}
\alias{test_arguments}
\title{Test (multiple) arguments of a prediction algorithm}
\usage{
test_arguments(fun, df_train, df_test, diagnostic_fun, arguments)
}
\arguments{
\item{fun}{prediction function}

\item{df_train}{training data}

\item{df_test}{testing data}

\item{diagnostic_fun}{the criteria with which the predictive performance will
be assessed}

\item{arguments}{named list of arguments to check}
}
\value{
a data.frame whose columns contain the diagnostics and run time, and
each row corresponds to a combination of the provided arguments
}
\description{
Test the performance of a prediction algorithm over a range of argument
values. Multiple arguments can be tested simultaneously.
}
\details{
\code{fun} should have formal arguments \code{df_train} and \code{df_test},
which are data used to train the model and test out-of-sample predictive
performance, respectively, as well as any arguments which are to be tested.
Each argument should require simple values only (a single number, string, etc.).
The value of \code{fun} should be a matrix-like object
with named columns and the same number of rows as \code{df_test}. The output
of \code{fun} is \code{cbind} to \code{df_test}, which is then passed into
\code{diagnostic_fun} to compute the diagnostics. Hence, since the number of
columns in the returned value of \code{fun} is arbitrary, one can test both
predictions and prediction uncertainty (e.g., by including prediction
standard errors or predictive interval bounds in the returned value of \code{fun}).
}
\seealso{
\code{\link{plot_diagnostics}}
}
