% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testargs_class.R
\docType{class}
\name{testargs-class}
\alias{testargs-class}
\title{testargs}
\description{
testargs
}
\section{Slots}{

\describe{
\item{\code{diagnostics_df}}{a data.frame containing the diagnostics for each combination of the supplied arguments}

\item{\code{arg_names}}{the argument names}

\item{\code{diagnostic_names}}{the diagnostic names}
}}

